Added example in demos for tag content reader
